+++ src/components/Employee/PaymentMethod/Head.tsx
@@ -1,5 +1,5 @@
 import { useTranslation } from 'react-i18next'
-import { usePaymentMethod } from '@/components/Employee/PaymentMethod/PaymentMethod'
+import { usePaymentMethod } from './usePaymentMethod'
 
 export function Head() {
   const { t } = useTranslation('Employee.PaymentMethod')
