+++ src/components/Employee/Taxes/StateForm.tsx
@@ -3,7 +3,7 @@ import { useFormContext, type Control } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
 import * as v from 'valibot'
 import type { EmployeeStateTaxQuestion } from '@gusto/embedded-api/models/components/employeestatetaxquestion'
-import { useTaxes } from './Taxes'
+import { useTaxes } from './useTaxes'
 import { TaxInputs } from '@/components/Common'
 import type { STATES_ABBR } from '@/shared/constants'
 import { snakeCaseToCamelCase } from '@/helpers/formattedStrings'
